+++ src/client/components/Footer/index.js
@@ -1,4 +1,6 @@
 import React from 'react'
+import Emoji from '../Emoji'
+import config from '../../config'
 import './index.css'
 
 const Footer = ({ text, className }) => (
@@ -6,6 +8,12 @@ const Footer = ({ text, className }) => (
     className={className}
   >
     {text}
+    <a href={config.coffeeUrl} target='_blank' rel='noopener noreferrer'>
+      <span>
+        Buy me a coffee&nbsp;
+        <Emoji emoji={'☕'} label='coffee' />
+      </span>
+    </a>
   </footer>
 )