+++ src/server/index.js
@@ -3,9 +3,12 @@ import express from 'express'
 import http from 'http'
 import path from 'path'
 import cors from 'cors'
+import config from 'config'
+import cronSchedule from 'node-schedule'
+import healthCheck from './utils/health-check'
 
 const app = express()
-const port = 9000
+const port = config.port
 
 app.use(cors())
 app.use(express.static(path.join(__dirname, '/../../build')))
@@ -18,6 +21,17 @@ const server = async () => {
   })
   httpServer.listen(port, () => {
     console.log(`Inspire me app listening on port ${port}!`)
+     // cron health check
+     const { endpoint, schedule, enabled } = config.healthCheck
+     if (enabled) {
+       console.info(`Health check is enabled for ${schedule}`)
+       cronSchedule.scheduleJob(
+         schedule,
+         async () => { await healthCheck(endpoint) }
+       )
+     } else {
+       console.warn('Health check is disabled')
+     }
   })
 }
 server()